Moved GraphPage to its own file
